HTTP Requester Cleanup
