+++ Sources/Network/Connection/Connections/ConsensusConnection.swift
@@ -46,7 +46,7 @@ final class ConsensusConnection:
                     }
                     return .http(httpService: ConsensusHttpConnection(
                                     config: config,
-                                    requester: RestApiRequester(requester: requester, baseUrl:config.url.httpBasedUrl),
+                                    requester: RestApiRequester(requester: requester, baseUrl: config.url.httpBasedUrl),
                                     targetQueue: targetQueue,
                                     rng: rng,
                                     rngContext: rngContext))
