+++ Sources/Network/Connection/Connections/FogBlockConnection.swift
@@ -38,7 +38,7 @@ final class FogBlockConnection:
                     }
                     return .http(httpService: FogBlockHttpConnection(
                                     config: config,
-                                    requester: RestApiRequester(requester: requester, baseUrl:config.url.httpBasedUrl),
+                                    requester: RestApiRequester(requester: requester, baseUrl: config.url.httpBasedUrl),
                                     targetQueue: targetQueue))
                 }
             },
