+++ Sources/Network/Connection/Connections/FogMerkleProofConnection.swift
@@ -46,7 +46,7 @@ final class FogMerkleProofConnection:
                     }
                     return .http(httpService: FogMerkleProofHttpConnection(
                                     config: config,
-                                    requester: RestApiRequester(requester: requester, baseUrl:config.url.httpBasedUrl),
+                                    requester: RestApiRequester(requester: requester, baseUrl: config.url.httpBasedUrl),
                                     targetQueue: targetQueue,
                                     rng: rng,
                                     rngContext: rngContext))
