+++ Sources/Network/Connection/Connections/FogUntrustedTxOutConnection.swift
@@ -39,7 +39,7 @@ final class FogUntrustedTxOutConnection:
                     }
                     return .http(httpService: FogUntrustedTxOutHttpConnection(
                                     config: config,
-                                    requester: RestApiRequester(requester: requester, baseUrl:config.url.httpBasedUrl),
+                                    requester: RestApiRequester(requester: requester, baseUrl: config.url.httpBasedUrl),
                                     targetQueue: targetQueue))
                 }
             },
