+++ Sources/Network/Connection/Connections/FogViewConnection.swift
@@ -46,7 +46,7 @@ final class FogViewConnection:
                     return .http(
                         httpService: FogViewHttpConnection(
                             config: config,
-                            requester: RestApiRequester(requester: requester, baseUrl:config.url.httpBasedUrl),
+                            requester: RestApiRequester(requester: requester, baseUrl: config.url.httpBasedUrl),
                             targetQueue: targetQueue,
                             rng: rng,
                             rngContext: rngContext))
