--- Sources/Network/HttpConnection/HTTPInterface/HTTP.swift
@@ -1,11 +1,0 @@
-//
-//  Copyright (c) 2020-2021 MobileCoin. All rights reserved.
-//
-
-import Foundation
-import SwiftProtobuf
-
-
-public enum HTTP {
-//    public typealias CallOptions = URLSessionConfiguration
-}
