+++ Sources/Network/HttpConnection/HTTPInterface/HTTPCallOptions.swift
@@ -4,13 +4,12 @@
 
 import Foundation
 
-
 public struct HTTPCallOptions {
     var headers: [String : String]
-    var timeoutIntervalForRequest : TimeInterval? = nil
-    var timeoutIntervalForResource : TimeInterval? = nil
-    
+    var timeoutIntervalForRequest : TimeInterval?
+    var timeoutIntervalForResource : TimeInterval?
 }
+
 extension HTTPCallOptions {
     public init() {
         self.init(headers: [:], timeoutIntervalForRequest: 30, timeoutIntervalForResource: 30)
