+++ Sources/Network/HttpConnection/HTTPInterface/HTTPMethod.swift
@@ -1,0 +1,14 @@
+//
+//  Copyright (c) 2020-2021 MobileCoin. All rights reserved.
+//
+
+import Foundation
+
+public enum HTTPMethod : String {
+    case GET = "GET"
+    case POST = "POST"
+    case PUT = "PUT"
+    case HEAD = "HEAD"
+    case PATCH = "PATCH"
+    case DELETE = "DELETE"
+}
